Move `struct Stacked` into own private module
